fix(static-analysis): fix 4 cppcheck violations + broken grep pattern
static_analysis.sh: grep pattern '\[.*\]:' never matched cppcheck 2.x
output (which ends with '[id]' not '[id]:'), so the script always reported
PASS even with violations present. Fixed to '\[.*\]'.

Violations fixed:
- comm_init.c: %u with signed int (GN_ADDRESS/OBC_ADDRESS) -> %d
- flight_mode_manager.c: cppcheck-suppress knownConditionTrueFalse on
  fault_get_highest_level() (weak stub returns 0 at analysis time)
- attitude_control_task.c: %llu with uint64_t (unsigned long on LP64) ->
  explicit (unsigned long long) cast
- sensor_read_task.c: same %llu / uint64_t cast fix

+++ src/tasks/attitude_control_task.c
@@ -107,7 +107,8 @@ void vAttitudeControlTask(void *pvParameters)
     if (samples >= 201)
     {  // 200 samples for 20Hz (~10 seconds)
       printf("[attitude_control_task] Timing (200 samples): min=%llu, max=%llu, avg=%llu us\n",
-             min_int, max_int, sum_int / 200);
+             (unsigned long long)min_int, (unsigned long long)max_int,
+             (unsigned long long)(sum_int / 200));
       min_int = 0xFFFFFFFFFFFFFFFF;
       max_int = 0;
       sum_int = 0;
